+++ .cargo/config.toml
@@ -2,4 +2,3 @@
 [build]
 rustflags = [ "--cfg=web_sys_unstable_apis" ]
 rustdocflags = [ "--cfg=web_sys_unstable_apis" ]
-target = "wasm32-unknown-unknown"
