+++ Cargo.toml
@@ -42,3 +42,50 @@ anyhow = "1.0.79"
 js-sys = "0.3.64"
 wasm-bindgen = "0.2.91"
 wasm-bindgen-test = "0.3.34"
+cfg-if = "1.0.0"
+chrono = "0.4.35"
+clap = "4.5.3"
+console_error_panic_hook = "0.1.7"
+console_log = "1.0.0"
+dot3 = "0.1.0"
+encase = { git = "https://github.com/cwfitzgerald/encase", branch = "add-member" }
+env_logger = "0.11.3"
+fern = "0.6.2"
+getrandom = "0.2"
+glam = "0.27.0"
+globwalk = "0.8.1"
+gloo-net = { version = "0.5.0", default-features = false }
+hound = "3.5.1"
+indexed_db_futures = "0.4.1"
+itertools = "0.12.1"
+lazy_static = "1.4.0"
+ndarray = "0.15.6"
+ndarray-stats = "0.5.1"
+num = "0.4.1"
+numpy = "0.20.0"
+parking_lot = "0.12.1"
+pathdiff = "0.2.1"
+pollster = "0.3.0"
+proptest = "1.4.0"
+pyo3 = "0.20.2"
+rand = "0.8.4"
+rand_distr = "0.4.3"
+realfft = "3.3.0"
+regex = "1.10.3"
+rustc-hash = "1.1.0"
+serde-wasm-bindgen = "0.6.5"
+serde_bytes = "0.11.14"
+serde_json = "1.0.114"
+slotmap = "1.0.7"
+smallvec = "1.11.2"
+strum = "0.25"
+strum_macros = "0.25"
+tabled = "0.15.0"
+tempfile = "3.3.0"
+tera = "1.19.0"
+test-strategy = "0.3.1"
+tokio = "1.36.0"
+uuid = "1.5.0"
+wasm-bindgen-futures = "0.4.41"
+web-sys = "0.3.64"
+web-time = "1.0.0"
