+++ crates/ratchet-cli/Cargo.toml
@@ -14,13 +14,13 @@ ratchet-models = { path = "../ratchet-models" }
 ratchet-hub = { path = "../ratchet-hub" }
 ratchet-nn = { path = "../ratchet-nn" }
 log.workspace = true
-clap = { version = "4.5.3", features = ["derive"] }
+clap = { workspace = true, features = ["derive"] }
 hf-hub = { workspace = true }
-serde_json = "1.0.114"
-env_logger = "0.11.3"
-fern = "0.6.2"
-chrono = "0.4.35"
+serde_json = { workspace = true }
+env_logger = { workspace = true }
+fern = { workspace = true }
+chrono = { workspace = true }
 tokenizers = "0.15.2"
-ndarray = "0.15.6"
-ndarray-stats = "0.5.1"
+ndarray = { workspace = true }
+ndarray-stats = { workspace = true }
 anyhow.workspace = true
