+++ crates/ratchet-core/src/ops/matmul.rs
@@ -251,24 +251,6 @@ impl GEMM {
             panic!("Transposed quantized inputs are not supported");
         }
 
-        //println!(
-        //    "MATMUL PROBLEM:
-        //lhs: {:?} {:?}
-        //rhs: {:?} {:?}
-        //bias: {:?}
-        //trans_lhs: {}
-        //trans_rhs: {}
-        //trans_out: {}",
-        //    lhs.shape(),
-        //    lhs.dt(),
-        //    rhs.shape(),
-        //    rhs.dt(),
-        //    bias.is_some(),
-        //    trans_lhs,
-        //    trans_rhs,
-        //    trans_out
-        //);
-
         Self {
             lhs,
             rhs,
@@ -530,7 +512,7 @@ impl MetaOperation for GEMM {
         let spec = self.compute_spec(dst);
 
         if spec.rhs_shape().is_vector() && !self.trans_lhs {
-            let (TILE_X, YT) = spec.heuristic.as_workload();
+            let (TILE_X, _) = spec.heuristic.as_workload();
             let group_x = WorkgroupCount::div_ceil(spec.lhs_shape()[0], TILE_X);
             let wgc = wgc![group_x as _, 1, spec.stacks() as _];
             Ok(wgc)
