+++ crates/ratchet-models/src/phi2/mod.rs
@@ -2,5 +2,5 @@ mod attn;
 mod mlp;
 mod model;
 
-pub use model::infer;
 pub use model::Phi2;
+pub mod generate;
