+++ crates/ratchet-models/src/phi2/model.rs
@@ -1,22 +1,19 @@
 use std::io::{BufRead, Seek};
 
-use crate::ratchet_from_gguf_web;
 use ratchet::{shape, Device, Tensor};
 use ratchet_loader::gguf::gguf::Header;
 use ratchet_nn::{Embedding, KVCache, KVEntry, LayerNorm, Linear, Module};
-use tokenizers::Tokenizer;
 
 use super::{
     attn::{PhiAttnInput, PhiSelfAttention},
     mlp::MLP,
 };
-use ndarray::Axis;
-use ndarray_stats::QuantileExt;
 
 #[cfg(target_arch = "wasm32")]
 use {
-    crate::TensorMap, js_sys::Uint8Array, ratchet_loader::gguf::gguf::ratchet_from_gguf,
-    std::collections::HashMap, wasm_bindgen::prelude::*,
+    crate::ratchet_from_gguf_web, crate::TensorMap, js_sys::Uint8Array,
+    ratchet_loader::gguf::gguf::ratchet_from_gguf, std::collections::HashMap,
+    wasm_bindgen::prelude::*,
 };
 
 #[derive(Debug)]
@@ -242,48 +239,6 @@ impl Phi2 {
     }
 }
 
-#[cfg(target_arch = "wasm32")]
-pub async fn infer(model: &mut Phi2, tokenizer: Tokenizer, prompt: String) -> anyhow::Result<()> {
-    use web_time::Instant;
-    log::warn!("Prompt: {}", prompt);
-    let encoding = tokenizer.encode(prompt, true).unwrap();
-    let mut tokens = encoding
-        .get_ids()
-        .iter()
-        .map(|&x| x as i32)
-        .collect::<Vec<_>>();
-    let mut all_logits = vec![];
-    let mut all_tokens = tokens.clone();
-    let mut loop_cnt = 0;
-    let start = Instant::now();
-    while tokens[tokens.len() - 1] != 50256 && loop_cnt < 256 {
-        let input = Tensor::from_data(
-            tokens.clone(),
-            shape![1, tokens.len()],
-            model.device.clone(),
-        );
-        let result = model.schedule(input)?.resolve()?;
-        let logits = result.to(&Device::CPU).await?;
-        all_logits.push(logits.clone());
-        model.cache_mut().update(tokens.len());
-
-        tokens = logits
-            .to_ndarray_view::<f32>()
-            .map_axis(Axis(2), |row| row.argmax_skipnan().unwrap())
-            .iter()
-            .map(|&x| x as i32)
-            .collect::<Vec<_>>();
-        let u32_toks = tokens.iter().map(|&x| x as u32).collect::<Vec<_>>();
-        log::warn!("{}", tokenizer.decode(&u32_toks, true).unwrap());
-        all_tokens.extend(tokens.clone());
-        loop_cnt += 1;
-    }
-    let elapsed = start.elapsed();
-    log::warn!("Elapsed: {:?}", elapsed);
-    log::warn!("Tok/s {}", all_tokens.len() as f64 / elapsed.as_secs_f64());
-    Ok(())
-}
-
 #[cfg(all(test, not(target_arch = "wasm32"), feature = "pyo3"))]
 mod tests {
     use hf_hub::api::sync::Api;
@@ -327,12 +282,12 @@ def ground():
         let _ = env_logger::builder().is_test(true).try_init();
         let api = Api::new().unwrap();
         let model_repo = api.model("FL33TW00D-HF/phi2".to_string());
-        let model_path = model_repo.get("phi2-q8_0.gguf").unwrap();
+        let model_path = model_repo.get("phi2-f16.gguf").unwrap();
         println!("MODEL PATH: {}", model_path.display());
 
         let mut reader = std::io::BufReader::new(std::fs::File::open(model_path)?);
         let device = Device::request_device(DeviceRequest::GPU)?;
-        let content = gguf::gguf::Content::read(&mut reader)?;
+        let content = gguf::gguf::Header::read(&mut reader)?;
         let mut model = Phi2::load(content, &mut reader, &device)?;
 
         let tokenizer_repo = api.model("microsoft/phi-2".to_string());
