+++ crates/ratchet-models/src/registry.rs
@@ -34,7 +34,6 @@ pub enum Whisper {
 #[cfg_attr(not(target_arch = "wasm32"), derive(clap::ValueEnum))]
 pub enum Phi {
     Phi2,
-    NanoLlama,
 }
 
 /// # Available Models
@@ -67,7 +66,6 @@ impl AvailableModels {
             },
             AvailableModels::Phi(p) => match p {
                 Phi::Phi2 => "FL33TW00D-HF/phi2",
-                Phi::NanoLlama => "jantxu/nano-llama", //TODO: remove just testing
             },
 
             _ => unimplemented!(),
@@ -88,7 +86,6 @@ impl AvailableModels {
             },
             AvailableModels::Phi(p) => match p {
                 Phi::Phi2 => "phi2",
-                Phi::NanoLlama => "ggml-model-f32.gguf",
             },
             _ => unimplemented!(),
         };
