+++ crates/ratchet-nn/src/embedding.rs
@@ -66,12 +66,8 @@ mod tests {
         }
     }
 
-    fn ground_truth(weight: &Tensor, indices: &Tensor, transposed: bool) -> anyhow::Result<Tensor> {
-        let arg = if transposed {
-            "torch.from_numpy(weight).t().contiguous()"
-        } else {
-            "torch.from_numpy(weight)"
-        };
+    fn ground_truth(weight: &Tensor, indices: &Tensor) -> anyhow::Result<Tensor> {
+        let arg = "torch.from_numpy(weight)";
 
         let prg = format!(
             r#"
@@ -94,14 +90,13 @@ def embedding(weight, indices):
         } = problem;
         let weight = Tensor::randn::<f32>(vocab_shape, Device::CPU);
 
-        let transposed = false;
-        let ground_truth = ground_truth(&weight, &indices, transposed).unwrap();
+        let ground_truth = ground_truth(&weight, &indices).unwrap();
         println!("GROUND TRUTH: {:?}", ground_truth);
 
         let weight = weight.to(&device).unwrap();
         let indices = indices.to(&device).unwrap();
 
-        let embedding = Embedding::new(weight, transposed);
+        let embedding = Embedding::new(weight);
         let result = embedding.schedule(indices).unwrap().resolve().unwrap();
         let x = result.to(&Device::CPU).unwrap();
         println!("OURS: {:?}", x);
@@ -126,49 +121,5 @@ def embedding(weight, indices):
     #[proptest(cases = 16)]
     fn test_embedding(prob: EmbeddingProblem) {
         run_embedding_trial(prob);
-    }
-
-    #[test]
-    fn dbg_phi_embedding() -> anyhow::Result<()> {
-        let api = Api::new().unwrap();
-        let model_repo = api.model("FL33TW00D-HF/phi2".to_string());
-        let model_path = model_repo.get("phi2-q8_0.gguf").unwrap();
-        //let model_path = model_repo.get("phi2-f16.gguf").unwrap();
-
-        println!("MODEL PATH: {}", model_path.display());
-
-        let mut reader = std::io::BufReader::new(std::fs::File::open(model_path)?);
-        let device = Device::request_device(DeviceRequest::GPU)?;
-        let content = Content::read(&mut reader)?;
-
-        let token_weight = content.tensor(&mut reader, "token_embd.weight", &Device::CPU)?;
-
-        let quantizer = Quantizer::new(Quantization::SInt8);
-        let dequant = quantizer.sint8_dequantize(token_weight);
-        println!("DEQUANT: {:?}", dequant.to_ndarray_view::<f32>());
-
-        let token_weight = content.tensor(&mut reader, "token_embd.weight", &Device::CPU)?;
-
-        //println!("TOKEN WEIGHT: {:?}", token_weight.to_ndarray_view::<f32>());
-        let embedding = Embedding::new(token_weight.to(&device)?, false);
-
-        let tokenizer_repo = api.model("microsoft/phi-2".to_string());
-        let tokenizer_path = tokenizer_repo.get("tokenizer.json").unwrap();
-        let tokenizer = Tokenizer::from_file(tokenizer_path).unwrap();
-
-        let prompt = "def print_prime(n):";
-        print!("{}", prompt);
-        let encoding = tokenizer.encode(prompt, true).unwrap();
-        let mut tokens = encoding
-            .get_ids()
-            .iter()
-            .map(|&x| x as i32)
-            .collect::<Vec<_>>();
-
-        let input = Tensor::from_data(tokens.clone(), shape![1, tokens.len()], device.clone());
-        let result = embedding.schedule(input).unwrap().resolve().unwrap();
-        let cpu_result = result.to(&Device::CPU).unwrap();
-        println!("CPU RESULT: {:?}", cpu_result.to_ndarray_view::<f32>());
-        Ok(())
     }
 }
