+++ crates/ratchet-web/Cargo.toml
@@ -26,26 +26,22 @@ wasm-opt = ['-O3', '--enable-simd']
 ratchet-models = { path = "../ratchet-models" }
 ratchet-hub = { path = "../ratchet-hub" }
 ratchet-loader = { path = "../ratchet-loader" }
-wasm-bindgen = "0.2.84"
-wasm-bindgen-futures = "0.4.39"
-js-sys = "0.3.64"
-indexed_db_futures = "0.4.1"
+wasm-bindgen = { workspace = true }
+wasm-bindgen-futures = { workspace = true }
+js-sys = { workspace = true }
+indexed_db_futures = { workspace = true }
 thiserror.workspace = true
 anyhow.workspace = true
-serde = "1.0.197"
-serde-wasm-bindgen = "0.6.5"
-console_error_panic_hook = "0.1.7"
-console_log = { version = "1.0.0", features=["color"] }
+serde = { workspace = true }
+serde-wasm-bindgen = { workspace = true }
+console_error_panic_hook = { workspace = true }
+console_log = { workspace = true, features = ["color"] }
 log.workspace = true
-hound = "3.5.1"
-fern = "0.6.2"
-chrono = "0.4.34"
-uuid = { version = "1.5.0", features=["v4","serde"] }
+hound = { workspace = true }
+fern = { workspace = true }
+chrono = { workspace = true }
+uuid = { workspace = true, features = ["v4", "serde"] }
 tokenizers = { version = "0.13.4", default-features = false, features=["unstable_wasm"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-getrandom = { version = "0.2.6", features = ["js"] }
-
 [dependencies.web-sys]
 features = [
   'console',
@@ -64,7 +60,11 @@ features = [
   'CacheStorage',
   'IdbKeyRange',
 ]
-version = "0.3.64"
+workspace = true
+
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+getrandom = { version = "0.2.6", features = ["js"] }
 
 [dev-dependencies]
 wasm-bindgen-test.workspace = true
